+++ src/Framework/Console/Console.php
@@ -16,8 +16,6 @@ use Lightpack\Console\Commands\CreateProvider;
 use Lightpack\Console\Commands\CreateRequest;
 use Lightpack\Console\Commands\CreateSeeder;
 use Lightpack\Console\Commands\CreateTool;
-use Lightpack\Console\Commands\DeployCommand;
-use Lightpack\Console\Commands\RollbackCommand;
 use Lightpack\Console\Commands\CreateTransformer;
 use Lightpack\Console\Commands\GenerateAppKey;
 use Lightpack\Console\Commands\LinkStorage;
@@ -30,6 +28,8 @@ use Lightpack\Console\Commands\SeedCommand;
 use Lightpack\Console\Commands\ServeCommand;
 use Lightpack\Console\Commands\UnlinkStorage;
 use Lightpack\Console\Commands\WatchCommand;
+use Lightpack\DevOps\Commands\DeployCommand;
+use Lightpack\DevOps\Commands\RollbackCommand;
 
 class Console
 {
